supervised_learning/transfer_learning/requirements-macos.txt
# Requirements spécifiques à macOS (Apple Silicon M1/M2)
# Optimisations GPU pour processeurs Apple Silicon
tensorflow-macos
tensorflow-metal
